azure-functions/requirements.txt
azure-functions
supabase>=2.7.4
httpx>=0.27.2
python-jose[cryptography]>=3.3.0
